[TASK] Rename commands
